refaktoring into structures & entities
I'm not good in explaining things with words, but maybe some code
examples show what I mean. Therefore this branch is intended as
discussion base, not as reference code.
Things I changed:
* Structures and Entities are now completely separated
* all handling for them is done in the separated classes (load and draw)
* Entities are handled with float coordinates and Structures as int
* Structure and Entitiy information is sotred only in mapview, not in
multiple places (not cempletely true for Entities)
* the Entity menu is fixed and has filter categories, the Structure menu
is still dynamicall build
* the properties view on doubleclick is disabled, as it was not working
for me from the beginning (to be fixed)
* the status bar information for Entities is broken (to be fixed)